[TASK] Migrate to use ES6 modules
Use new @typo3 scoped namespace
and convert to ES6 import-syntax.

Releases: main
Forge: https://forge.typo3.org/issues/97006
Gerrit: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73675
