+++ Configuration/JavaScriptModules.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+    'dependencies' => [
+        'core',
+        'backend',
+    ],
+    'imports' => [
+        '@typo3/styleguide/' => 'EXT:styleguide/Resources/Public/JavaScript/',
+    ],
+];
