+++ Resources/Private/Templates/Backend/Tab.html
@@ -9,8 +9,8 @@
 <f:section name="Content">
 
     <f:be.pageRenderer
-        includeRequireJsModules="{
-            0: 'TYPO3/CMS/Backend/Tabs'
+        includeJavaScriptModules="{
+            0: '@typo3/backend/tabs.js'
         }"
     />
